Move WebDriverCommandMsg to script_traits.
