+++ components/webdriver_server/lib.rs
@@ -32,13 +32,13 @@ use image::{DynamicImage, ImageFormat, RgbImage};
 use ipc_channel::ipc::{self, IpcReceiver, IpcSender};
 use keys::keycodes_to_keys;
 use msg::constellation_msg::{FrameId, LoadData, PipelineId};
-use msg::constellation_msg::{NavigationDirection, PixelFormat, WebDriverCommandMsg};
+use msg::constellation_msg::{NavigationDirection, PixelFormat};
 use msg::webdriver_msg::{LoadStatus, WebDriverCookieError, WebDriverFrameId};
 use msg::webdriver_msg::{WebDriverJSError, WebDriverJSResult, WebDriverScriptCommand};
 use regex::Captures;
 use rustc_serialize::base64::{CharacterSet, Config, Newline, ToBase64};
 use rustc_serialize::json::{Json, ToJson};
-use script_traits::ConstellationMsg;
+use script_traits::{ConstellationMsg, WebDriverCommandMsg};
 use std::borrow::ToOwned;
 use std::collections::BTreeMap;
 use std::net::{SocketAddr, SocketAddrV4};
